(UI) Hack AddDelayFeature
Since this add feature is custom, its behavior differs from other
add/create feature: it both add the graphical element and create
the business object. In order to programmatically use it and keep
consistent links between pictogram elements and business object,
one has to overwrite the links after the feature. This hack add a
list to keep track of the created pictogram elements.