feat: Add co-iteration of SrSDF using DAG
Traverse a single rate SDF graph using the instances from
DAG and using instances of DAG as reference. The innovation is
that DAG and SrSDF are similar but not identical. Several implode
explode instances are not present in DAG and may be present in
SrSDF. We traverse SrSDF such that a subset of nodes of DAG is
mandatorily seen while not exiting when an unknown node is encountered.

Co-Iterator is needed while creating transient and cyclic DAG