feat (BREAKING): Change DAG generator to include subgraphs
Strongly connected graphs are subgraphs. DAG generation is modified
to be compatible with subgraphs as well

The check to ensure input is valid is no longer consistently used.
This is delegated to be a private method when needed. The SDF needs
to be checked for consistency BEFORE passing it to DAG generator.

Modify tests accordingly