fix: Generation of DAG works for graphs with special actors
When subgraphs are converted to DAGs, special actors such as
broadcasts can be left dangling i.e. some of its output ports
will be left unconnected. SDF2DAG flags this as an error. As
dangling ports cannot be avoided, we add a check to allow it only
for subgraphs.

Similarly, DAGUtils only checks names for coiteration as port size and
names may not be equal due to dangling ports. The test is
updated to reflect the same.