fix: Properly consider neighbours of interface nodes
Interface nodes connects a strongly connected component (SCC) with rest of the
graph. The neighbours of interface nodes can prevent rearranging of the
component although the SCC is instance independent. The fix checks for
this condition and reports the failure appropriately.

This means, instead of co-iterating from root nodes, we need to
topologically co-iterate the entire DAG of a SCC.

Tests are updated to reflect the changes.

Minor bug correction in SDF2DAG

Support elaborate Debug messages

Perform SCC detection on original SDF instead of Acyclic-like graph

Correct false reports of DAG as instance dependent

Overload SDF export class