fix: Upgrade to new version of jgraph lib
Remove implementation of SCC detection as new version has it already
Change deprecated DirectedSubgraph to AsSubgraph
Sort the manual arrays before comparison
Check if the edge exists before its removal (AcyclicLikeSubgraphDetector).

There could be a potential bug in the DFTools/jgrapht library as failure to check
this results in stackoverflow. Issue rasied from an exception which is supposed
to inform about the missing edge. But it requires name of the edge, which
triggers stack overflow.